+++ src/main/scala/ai/starlake/schema/generator/Yml2Xls.scala
@@ -1,10 +1,10 @@
 package ai.starlake.schema.generator
 
 import ai.starlake.config.{DatasetArea, Settings}
-import ai.starlake.schema.handlers.SchemaHandler
+import ai.starlake.schema.handlers.{SchemaHandler, StorageHandler}
 import ai.starlake.schema.model._
-import better.files.File
 import com.typesafe.scalalogging.LazyLogging
+import org.apache.hadoop.fs.Path
 import org.apache.poi.ss.usermodel.CellType
 import org.apache.poi.xssf.usermodel.XSSFWorkbook
 
@@ -33,7 +33,7 @@ class Yml2Xls(schemaHandler: SchemaHandler) extends LazyLogging with XlsModel {
     }
 
     domains.foreach { domain =>
-      writeDomainXls(domain, outputDir)
+      writeDomainXls(domain, outputDir)(settings.storageHandler())
     }
 
     schemaHandler
@@ -63,10 +63,12 @@ class Yml2Xls(schemaHandler: SchemaHandler) extends LazyLogging with XlsModel {
     }
   }
 
-  def writeDomainXls(domain: Domain, folder: String): Unit = {
+  def writeDomainXls(domain: Domain, folder: String)(implicit
+    storageHandler: StorageHandler
+  ): Unit = {
     val workbook = new XSSFWorkbook()
 
-    val xlsOut = File(folder, domain.name + ".xlsx")
+    val xlsOut = new Path(folder, domain.name + ".xlsx")
     val domainSheet = workbook.createSheet("_domain")
     fillHeaders(workbook, allDomainHeaders, domainSheet)
     val domainRow = domainSheet.createRow(2)
@@ -197,9 +199,12 @@ class Yml2Xls(schemaHandler: SchemaHandler) extends LazyLogging with XlsModel {
         attributesSheet.autoSizeColumn(i)
 
     }
-
-    xlsOut.delete(swallowIOExceptions = true)
-    workbook.write(xlsOut.newFileOutputStream(false))
+    val outputStream = storageHandler.output(xlsOut)
+    try {
+      workbook.write(outputStream)
+    } finally {
+      outputStream.close()
+    }
   }
 
 }